+++ earthaccess/store.py
@@ -1,4 +1,5 @@
 import datetime
+import logging
 import shutil
 import traceback
 from functools import lru_cache
@@ -20,6 +21,8 @@ from .auth import Auth
 from .daac import DAAC_TEST_URLS, find_provider
 from .results import DataGranule
 from .search import DataCollections
+
+logger = logging.getLogger(__name__)
 
 
 class EarthAccessFile(fsspec.spec.AbstractBufferedFile):
@@ -110,7 +113,7 @@ class Store(object):
                     self.set_requests_session(url)
 
         else:
-            print("Warning: the current session is not authenticated with NASA")
+            logger.warn("The current session is not authenticated with NASA")
             self.auth = None
         self.in_region = self._running_in_us_west_2()
 
@@ -339,7 +342,7 @@ class Store(object):
     ) -> List[Any]:
         fileset: List = []
         total_size = round(sum([granule.size() for granule in granules]) / 1024, 2)
-        print(f"Opening {len(granules)} granules, approx size: {total_size} GB")
+        logger.info(f"Opening {len(granules)} granules, approx size: {total_size} GB")
 
         if self.auth is None:
             raise ValueError(
@@ -353,10 +356,10 @@ class Store(object):
                 # if the data has its own S3 credentials endpoint, we will use it
                 endpoint = self._own_s3_credentials(granules[0]["umm"]["RelatedUrls"])
                 if endpoint is not None:
-                    print(f"using endpoint: {endpoint}")
+                    logger.info(f"using endpoint: {endpoint}")
                     s3_fs = self.get_s3fs_session(endpoint=endpoint)
                 else:
-                    print(f"using provider: {provider}")
+                    logger.info(f"using provider: {provider}")
                     s3_fs = self.get_s3fs_session(provider=provider)
             else:
                 access = "on_prem"
@@ -425,7 +428,7 @@ class Store(object):
                             f"Exception: {traceback.format_exc()}"
                         ) from e
                 else:
-                    print(f"Provider {provider} has no valid cloud credentials")
+                    logger.info(f"Provider {provider} has no valid cloud credentials")
                 return fileset
             else:
                 raise ValueError(
@@ -523,13 +526,13 @@ class Store(object):
                 "we need to use one from earthaccess.list_cloud_providers()"
             )
         if self.in_region and data_links[0].startswith("s3"):
-            print(f"Accessing cloud dataset using provider: {provider}")
+            logger.info(f"Accessing cloud dataset using provider: {provider}")
             s3_fs = self.get_s3fs_session(provider=provider)
             # TODO: make this parallel or concurrent
             for file in data_links:
                 s3_fs.get(file, str(local_path))
                 file_name = local_path / Path(file).name
-                print(f"Downloaded: {file_name}")
+                logger.info(f"Downloaded: {file_name}")
                 downloaded_files.append(file_name)
             return downloaded_files
 
@@ -559,17 +562,17 @@ class Store(object):
             )
         )
         total_size = round(sum([granule.size() for granule in granules]) / 1024, 2)
-        print(
+        logger.info(
             f" Getting {len(granules)} granules, approx download size: {total_size} GB"
         )
         if access == "direct":
             if endpoint is not None:
-                print(
+                logger.info(
                     f"Accessing cloud dataset using dataset endpoint credentials: {endpoint}"
                 )
                 s3_fs = self.get_s3fs_session(endpoint=endpoint)
             else:
-                print(f"Accessing cloud dataset using provider: {provider}")
+                logger.info(f"Accessing cloud dataset using provider: {provider}")
                 s3_fs = self.get_s3fs_session(provider=provider)
 
             local_path.mkdir(parents=True, exist_ok=True)
@@ -578,7 +581,7 @@ class Store(object):
             for file in data_links:
                 s3_fs.get(file, str(local_path))
                 file_name = local_path / Path(file).name
-                print(f"Downloaded: {file_name}")
+                logger.info(f"Downloaded: {file_name}")
                 downloaded_files.append(file_name)
             return downloaded_files
         else:
@@ -615,11 +618,10 @@ class Store(object):
                         # https://docs.python-requests.org/en/latest/user/quickstart/#raw-response-content
                         shutil.copyfileobj(r.raw, f, length=1024 * 1024)
             except Exception:
-                print(f"Error while downloading the file {local_filename}")
-                print(traceback.format_exc())
+                logger.exception(f"Error while downloading the file {local_filename}")
                 raise Exception
         else:
-            print(f"File {local_filename} already downloaded")
+            logger.info(f"File {local_filename} already downloaded")
         return str(path)
 
     def _download_onprem_granules(
@@ -663,8 +665,7 @@ class Store(object):
             try:
                 fileset = _open_files(url_mapping, https_fs, threads)
             except Exception:
-                print(
-                    "An exception occurred while trying to access remote files via HTTPS: "
-                    f"Exception: {traceback.format_exc()}"
+                logger.exception(
+                    "An exception occurred while trying to access remote files via HTTPS"
                 )
         return fileset